fix(api): Changed isPending to Incomplete in adding blank rental
